+++ pbge/dialogue/__init__.py
@@ -302,7 +302,7 @@ def list_nouns(mylist, conjunction="and"):
     if len(mywords) > 2:
         return ", ".join(mywords[:-1]) + ", {} {}".format(conjunction, mywords[-1])
     elif len(mywords) > 1:
-        return "{} {} {}".join(mywords[0], conjunction, mywords[1])
+        return "{} {} {}".format(mywords[0], conjunction, mywords[1])
     elif mywords:
         return mywords[0]
     else: